refactor codegen files
