+++ mithril/framework/codegen/py_style_codegen/python_gen.py
@@ -21,22 +21,22 @@ from functools import partial
 from posixpath import basename, splitext
 from typing import Any, Generic, Protocol
 
-from ...backends.backend import ParallelBackend
-from ...common import PythonGenConfig
-from ...types import DataType, Dtype
-from ...utils.func_utils import prepare_function_args
-from ..common import (
+from ....backends.backend import ParallelBackend
+from ....common import PythonGenConfig
+from ....types import DataType, Dtype
+from ....utils.func_utils import prepare_function_args
+from ...common import (
     DataEvalType,
     EvaluateAllType,
     EvaluateType,
     FinalCost,
     ParamsEvalType,
 )
-from ..logical import Operator
-from ..physical.model import PhysicalModel
-from ..utils import GeneratedFunction
-from .code_gen import CodeGen
-from .utils import (
+from ...logical import Operator
+from ...physical.model import PhysicalModel
+from ...utils import GeneratedFunction
+from ..code_gen import CodeGen
+from ..utils import (
     convert_to_ast_arg,
     convert_to_ast_kwarg,
     partial_array_creation_func,
